refactor(cli): make cli more clean
