+++ metassr-cli/src/cli/mod.rs
@@ -1,7 +1,13 @@
-use std::{fmt::Display, str::FromStr};
+mod builder;
+mod creator;
+mod runner;
+pub mod traits;
+
+pub use builder::*;
+pub use runner::*;
 
 use clap::{command, Parser, Subcommand, ValueEnum};
-use metassr_build::server;
+
 #[derive(Parser, Debug)]
 #[command(
     author,
@@ -55,43 +61,4 @@ pub enum Commands {
         #[arg(long)]
         serve: bool,
     },
-}
-
-#[derive(Debug, ValueEnum, PartialEq, Eq, Clone)]
-pub enum BuildingType {
-    /// Static-Site Generation.
-    SSG,
-    /// Server-Side Rendering.
-    SSR,
-}
-
-impl Into<server::BuildingType> for BuildingType {
-    fn into(self) -> server::BuildingType {
-        match self {
-            Self::SSG => server::BuildingType::StaticSiteGeneration,
-            Self::SSR => server::BuildingType::ServerSideRendering,
-        }
-    }
-}
-
-impl Display for BuildingType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(match *self {
-            Self::SSG => "ssg",
-            Self::SSR => "ssr",
-        })
-    }
-}
-
-impl FromStr for BuildingType {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "ssr" => Ok(BuildingType::SSR),
-            "server-side rendering" => Ok(BuildingType::SSR),
-            "ssg" => Ok(BuildingType::SSG),
-            "static-site generation" => Ok(BuildingType::SSG),
-            _ => Err("unsupported option.".to_string()),
-        }
-    }
 }
