+++ metassr-cli/src/cli/traits.rs
@@ -1,0 +1,8 @@
+use anyhow::Result;
+
+pub trait Exec {
+    fn exec(&self) -> Result<()>;
+}
+pub trait AsyncExec {
+    async fn exec(&self) -> Result<()>;
+}
